Update difficulties
Instead of changing the values of the WinningNetWorth and LosingNetWorth variables, when choosing any of the difficulties, the PositiveEventChance variable changes. As the name implies, it changes the chance of a positive event occurring. This implementation is better as it really changes the difficulty of the game and not just how long the player will play.